Arduino/python/pyproject.toml
[tool.poetry]
name = "max30009-python"
version = "0.1.0"
description = ""
authors = ["void2610 <izumisan2610@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.0"
pyqtgraph = "^0.13.7"
pyserial = "^3.5"
PyAutoGUI = "^0.9.54"
PyQt5 = "^5.15.11"
pygame = "^2.6.0"
keyboard = "^0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
